1、表定义 2、完善租车套餐设计逻辑

+++ src/main/java/com/xiliulou/electricity/vo/car/CarRentalPackageVO.java
@@ -182,7 +182,7 @@ public class CarRentalPackageVO implements Serializable {
     /**
      * 创建时间
      */
-    private Integer createTime;
+    private Long createTime;
 
     // ++++++++++ 辅助业务数据 ++++++++++
